Adding k-fold-cross valication for training and plottong/evaluation
New function make-kfolds creates and saves indices for train and test sets for a specified number of folds
New parameters kfolds and nfolds specify which kfold to train on and the total number of folds.
For the parameter kfolds, the default is set to 0, meaning that cross-validation should not be used.
If the parameter is specified to a number >0 and <= nfolds, the specified fold data will be loaded and used for training.
A new function is added to create all basic plots with all kfolds, by concatenating the predictions, truth labers etc. This requires to specify the kfold set used by indicating the total number of folds